% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fonts.R
\name{check_fonts}
\alias{check_fonts}
\title{Check if the Proxima Nova font is installed}
\usage{
check_fonts()
}
\description{
This function checks if the Proxima Nova font is installed on the user's system. If the font is not installed, it provides instructions on how to set up the font for use in the Quarto project.
}
