% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fonts.R
\name{check_proxima_nova_installed}
\alias{check_proxima_nova_installed}
\title{Internal: Check if the Proxima Nova font is installed}
\usage{
check_proxima_nova_installed()
}
\value{
TRUE if the Proxima Nova font is installed, FALSE otherwise
}
\description{
This function checks if the Proxima Nova font is installed on the user's system.
}
