% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fonts.R
\name{display_font_setup_instructions}
\alias{display_font_setup_instructions}
\title{Internal: Function to display instructions for setting up the font}
\usage{
display_font_setup_instructions()
}
\description{
This function displays instructions for setting up the Proxima Nova font for use in the Quarto project.
}
