% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_fonts.R
\name{download_fonts}
\alias{download_fonts}
\title{Download Proxima Nova fonts from the AAGI private repository}
\usage{
download_fonts(
  font_repo = "jafernandez01/font-resources",
  font_dir = "fonts/",
  dest_dir = "./_extensions/aagi-report",
  pat = Sys.getenv("GITHUB_PAT")
)
}
\arguments{
\item{font_repo}{The GitHub repository containing the font files}

\item{font_dir}{The directory within the repository containing the font files}

\item{dest_dir}{The destination directory to extract the font files}

\item{pat}{The GitHub Personal Access Token (PAT) to access the private repository}
}
\value{
TRUE if the fonts are downloaded and extracted successfully, FALSE otherwise
}
\description{
This function downloads the Proxima Nova font files from the private AAGI repository and extracts them to the specified destination directory.
}
