% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_aagi_document.R
\name{new_aagi_document}
\alias{new_aagi_document}
\title{Create new Quarto document from template}
\usage{
new_aagi_document(
  file_name = NULL,
  ext_name = "aagi-report",
  university = "UQ",
  update = FALSE
)
}
\arguments{
\item{file_name}{Name of new qmd file and sub-directory to be created}

\item{ext_name}{String indicating which extension to install}

\item{university}{String indicating which university partner to use}

\item{update}{Logical indicating whether to update the extension if it already exists}
}
\description{
Create a new AAGI Quarto document. If the extension is not already
installed, it will be copied over to the \verb{_extensions} folder in the
current working directory. This function extends a function written by
Thomas Mock: https://github.com/jthomasmock/octavo/blob/master/R/use_quarto_ext.R
and by Spencer Schien: https://spencerschien.info/post/r_for_nonprofits/quarto_template/
}
