% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_project.R
\name{setup_docs}
\alias{setup_docs}
\title{Rename the index.qmd File and Update _quarto.yml for a Quarto Project}
\usage{
setup_docs(output_type = "Report")
}
\arguments{
\item{output_type}{Character. The type of output to be generated (e.g., "Report", "Slides", "Manuscript").}
}
\description{
This function renames the \code{index.qmd} file to a new name based on the project name and output type.
It also updates the \verb{_quarto.yml} file to specify the output type and directory for the project.
This function should be run in the root directory of a Quarto project and after the \code{setup_folders()} function.
}
\details{
The function extracts the project name from the directory name and uses it to rename the \code{index.qmd} file.
The new file name format is \verb{01_projectname_outputtype.qmd}.
The \verb{_quarto.yml} file is updated with the output type and directory for the project.
}
\examples{
# Rename the index.qmd file and update _quarto.yml for a project report
setup_docs("Report")

}
