% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_project.R
\name{setup_folders}
\alias{setup_folders}
\title{Create a Standard Folder Structure for a Quarto Project}
\usage{
setup_folders(project_path = ".")
}
\arguments{
\item{project_path}{Character. The root path of the Quarto project where the folders will be created.}
}
\description{
This function creates a standardized folder structure for a Quarto project.
The structure includes folders for data, code, figures, documents, outputs, source files, and admin tasks.
}
\details{
\itemize{
\item \strong{data}: Contains the raw data files used in the project. These files should not be altered and are ideally read-only.
\item \strong{code}: Contains the scripts or code files used to run the project.
\item \strong{figs}: Contains plots, images, tables, or figures created and saved by your code.
It should be possible to delete and regenerate this folder with the scripts in the project.
\item \strong{doc}: Contains any manuscripts or interim summaries produced during the project.
\item \strong{output}: Contains non-figure objects created by the scripts, such as processed data or logs.
\item \strong{src}: Contains any files you may want to \code{source()} in your scripts, such as .R files containing helper functions.
\item \strong{admin}: Contains administrative files related to the project (e.g., notes, project planning documents, and metadata).
}
}
\note{
Folder structure inspired by Kathryn Destasio's best practices for R projects: https://kdestasio.github.io/post/r_best_practices/
}
\examples{
# Create the folder structure in the current working directory
setup_folders(".")

}
