% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_fonts.R
\name{setup_github_pat}
\alias{setup_github_pat}
\title{Helper function to setup GITHUB_PAT interactively}
\usage{
setup_github_pat()
}
\value{
TRUE if the GITHUB_PAT is successfully set up, FALSE otherwise
}
\description{
This function provides instructions to the user on how to create a GitHub Personal Access Token (PAT) and set it up for use in the current R session.
}
